TESTING

+++ scripts/stop_container.sh
@@ -1,8 +1,13 @@
-#!/bin/bash
-set -e
+
+
+# Install Docker if it's not already installed
+if ! [ -x "$(command -v docker)" ]; then
+  echo 'Error: Docker is not installed.' >&2
+  exit 1
+fi
 
 # Stop the running container (if any)
-containerid = 'docker ps | awk '{print $1}'
-docker rm -f  $containerid
+containerid=$(docker ps | awk '{print $1}')
+docker rm -f $containerid
